Add builder option to enable clicking though transparent pixels on images
- Added an option in PhotoEditor to enable clicking through transparent pixels.
- Added logic in MultiTouchListener to check event click location for transparent pixel and return false if pixel is transparent (e.g. click goes through to next layer)

This functionality enables the possibility of moving images that are below other images as long as they are visible. It makes it a lot easier to select and move an image among a group of other images (where you before had to basically move the top image away to be able to select the image below it).

The feature currently only works on images, but it could be applied to text, emoji, ect with future updates.