Added radius check to transparentPixelClickThroughRadius
- Added a transparentPixelsClickThroughRadius variable to indicate a radius in pixels in which the transparentPixel method will check for transparent pixels.

The radius check is there to help developer fine tune the transparent clickthrough (so you may want to be able to click through transparent pixels but still want to hit text with a lot of transparency inbetween the letters)

- Created method in MultiTouchListener to calculate if the pixels (inside the radius) around the clicked pixel are transparent or opaque.

The method recursively checks adjacent pixels for transparency then move onto those pixels and checks their adjacent pixels. This goes on until bitmap edge, radius limit or opaque pixel is reached.