In this version i have added custom fonts.
It cycles through fonts as the font button is pressed, an alternative to this would be a drop down menu of fonts, i don't want to impact too heavily on UI changes so i have left it as a cycle.

variables that are used when editing the text in TextEditorDialogFragment.java had to be passed through EditImageActivity.java
to then go into PhotoEditor.java by overloading the constructors  .

all fonts work with shadow layer adjustments and Bold and Italic features.