Change default arrow pointer location
Change meaning of ArrowPointerLocation.START and
ArrowPointerLocation.END. START point is first touched point and END
is the last touched. This was inverted in initial commit.

After few days of testing I realised that it's easier to draw arrow
when pointer is placed on arrow's START, so I decided to change
default pointer location. Because START and END meaning was changed
in the commit, default pointer location value in ShapeType.Arrow()
don't have to be changed.

Also "pointer location" sounds better than "pointer position" so the
name was also changed.